:hammer: Simplify the screen code
Now that I've only got the one screen, and given it doesn't actually do
much (the bulk of the clever work is in the package information widget), it
makes sense to put that small amount of work on the app/main screen.
