--- pispy/screens/__init__.py
@@ -1,11 +1,0 @@
-"""The screens that make up the application."""
-
-##############################################################################
-# Local imports.
-from .lookup import Lookup
-
-##############################################################################
-# Export the screens.
-__all__ = ["Lookup"]
-
-### __init__.py ends here
